feat: implement PR feedback
